+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Fix the Signal integration to handle unexpected RPC messages better (#763)
 - Fix special character encoding in Signal notifications (#767)
 - Fix project sort order to be case-insensitive everywhere in the UI (#768)
+- Fix special character encoding in project invite emails
 
 ## v2.5 - 2022-12-14